This is from Matt LeBlanc. He says:
It had set_error returning undef. While this works properly for cases where its
return is evaluated in scalar context, this is less than ideal when we're
talking about list context. i.e.

# perl -le'sub a { return undef } @a = a() or print "Left side evaluated
as false"'
# perl -le'sub a { return } @a = a() or print "Left side evaluated as
false"'
Left side evaluated as false
#

This makes a difference for how the dbfile driver fails when it can't
tie the file (otherwise it trudges on and says can't use string "" as a
subroutine).